Update folder organization
