--- README.md
@@ -1,7 +1,0 @@
-# Checkers-Game
-- Created by: Alex Kanotz
-
--- Try it out:
-- Checkers.jar is a runnable executable JAR file.
-- Select Checkers.jar, download, double click, and enjoy!
-- NOTE: Java Runtime Environment needed prior to executing.
