--- projectMain.java
@@ -1,19 +1,0 @@
-// **********************************************************
-// * Name:    Alex Kanotz                                   *
-// * Project: Checkers Game                                 *
-// * Date:    12/6/22                                       *
-// **********************************************************
-
-
-// **********************************************************
-// *                   projectMain class                    *
-// **********************************************************
-public class projectMain {
-
-    // **********************************************************
-    // *                 projectMain constructor                *
-    // **********************************************************
-    public static void main(String[] args) {
-        new startMenu();
-    } // End projectMain constructor
-} // End projectMain class
