+++ src/projectMain.java
@@ -1,0 +1,18 @@
+// **********************************************************
+// * Name:    Alex Kanotz                                   *
+// * Project: Checkers Game                                 *
+// **********************************************************
+
+
+// **********************************************************
+// *                   projectMain class                    *
+// **********************************************************
+public class projectMain {
+
+    // **********************************************************
+    // *                       main method                      *
+    // **********************************************************
+    public static void main(String[] args) {
+        new startMenu();
+    } // End main method
+} // End projectMain class
